Merged SDK changes from the Valve Github repo up until commit: "Added many shader source files" (7309a5f13f63fdcc7b1e090f6c176113a9d95061)
Log comments from the Valve commits:

* Added many shader source files

This should include the latest version of every shader that was in the
2007 SDK. It also includes a smattering of debug shaders, both VR
distortion shaders, and other assorted shaders that will hopefully be
useful.

None of these new files are included in the game shader DLL project. If
you need to modify one of these shaders for use in your mod you will
need to rename it so that you don't collide with the version of that
shader that lives in stdshader_dx9.dll.